# reference/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Retail Identity Service API
  version: "1.0.0"
  description: |
    Identity & Company service for Retail Management System.

    Key concepts:
    - Public signup creates only a global user identity
    - Authenticated users register companies with POST /api/v1/companies
    - Users are global identities and may belong to multiple companies
    - JWTs are user-level and contain sub=userUuid and username; company access is checked against membership in the database
    - Company-scoped APIs use /api/v1/companies/{companyUuid}/...
    - Company membership roles are OWNER, ADMIN, and MEMBER
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Required UUID used to prevent duplicate POST requests
...
    CompanyUuidPath:
      name: companyUuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Company UUID
  schemas:
...
    UUID:
      type: string
      format: uuid
    DateTime:
      type: string
      format: date-time

    # ---------- AUTH ----------
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          format: email
          maxLength: 150
          example: user@example.com
        password:
          type: string
          format: password
          maxLength: 128
          example: StrongPass@123
      additionalProperties: false
    LoginResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - tokenType
        - user
      properties:
        accessToken:
          type: string
          description: 'User-level JWT with claims: sub=userUuid, username, exp'
        refreshToken:
          type: string
          maxLength: 256
        expiresIn:
          type: integer
          format: int32
          example: 3600
        tokenType:
          type: string
          default: Bearer
          example: Bearer
        user:
          $ref: '#/components/schemas/UserResponse'
      additionalProperties: false
...
    CreateCompanyUserRequest:
      type: object
      required:
        - username
        - companyRole
      properties:
        username:
          type: string
          format: email
          maxLength: 150
          example: cashier@example.com
        displayName:
          type: string
          maxLength: 255
          example: Store User
        companyRole:
          $ref: '#/components/schemas/CompanyRole'
      additionalProperties: false
...
    CompanyUsersResponse:
      type: object
      required:
        - users
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
      additionalProperties: false
    UserResponse:
      type: object
      required:
        - userUuid
        - username
        - status
        - createdAt
      properties:
        userUuid:
          allOf:
            - $ref: '#/components/schemas/UUID'
        username:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        roles:
          type: array
          description: Operational roles only; not used for company membership.
          items:
            type: string
        companyRole:
          allOf:
            - $ref: '#/components/schemas/CompanyRole'
          nullable: true
          description: Present only for company-scoped user responses.
        status:
          type: string
          enum: [pending, active, suspended]
        createdAt:
          $ref: '#/components/schemas/DateTime'
      additionalProperties: false
...
    CompanyRole:
      type: string
      enum: [OWNER, ADMIN, MEMBER]
      description: Company membership role. Operational job permissions are deferred.
...
    ErrorResponse:
      type: object
      required:
        - serviceCode
        - code
        - message
      properties:
        serviceCode:
          type: string
          example: RMS-101
        code:
          type: string
          example: 400-1-3
        message:
          type: string
          example: Request validation failed
        details:
          type: object
          nullable: true
          additionalProperties: true
      additionalProperties: false
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /api/v1/auth/login:
    post:
      security: []
      summary: Login and receive JWT + refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/companies/{companyUuid}/users:
    get:
      summary: List company users
      description: OWNER or ADMIN may list users in the company, including suspended memberships for management screens.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CompanyUuidPath'
      responses:
        '200':
          description: Company users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyUsersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Create a user inside a company (admin only)
      description: OWNER or ADMIN may create a pending company user and request email verification. Password setup/reset flow is handled separately.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CompanyUuidPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyUserRequest'
      responses:
        '201':
          description: User created in company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...