# src/RMS.Identity.Service.Api/Endpoint/Auth/Login/LoginController.cs
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.RateLimiting;
using RMS.Identity.Service.Domain.Contracts.Login;
using RMS.Identity.Service.Infrastructure.Abstractions.Cqrs;

namespace RMS.Identity.Service.Api.Endpoint.Auth.Login;

[ApiController]
[Route("api/v1/auth/login")]
public sealed class LoginController : ControllerBase
{
    private readonly ICommandHandler<LoginCommandRequest, LoginCommandResponse> _commandHandler;

    public LoginController(ICommandHandler<LoginCommandRequest, LoginCommandResponse> commandHandler)
    {
        _commandHandler = commandHandler;
    }

    [HttpPost]
    [EnableRateLimiting("auth")]
    [ProducesResponseType(typeof(LoginResponse), StatusCodes.Status200OK)]
    [ProducesResponseType(typeof(object), StatusCodes.Status400BadRequest)]
    [ProducesResponseType(typeof(object), StatusCodes.Status401Unauthorized)]
    [ProducesResponseType(typeof(object), StatusCodes.Status403Forbidden)]
    public async Task<IActionResult> PostAsync(LoginRequest request, CancellationToken cancellationToken)
    {
        var login = await _commandHandler.HandleAsync(request.ToCommand(), cancellationToken);
        return Ok(login.ToResponse());
    }
}